+++ client/js/gdpr/analysis.js
